+++ lib/modular_navigation.dart
@@ -19,7 +19,7 @@ part 'pages/noparameters_modular_page.dart';
 
 part 'navigator/route_information_parser.dart';
 part 'navigator/router_delegate.dart';
-part 'routes/modular_route.dart';
+
 part 'navigator/modular_history.dart';
 part 'navigator/modular_link.dart';
 
@@ -30,6 +30,10 @@ part 'parameters/optional_id_page_parameters.dart';
 
 part 'routes/noanimationpageroute.dart';
 part 'routes/nopageparameters_route.dart';
+part 'routes/base_modular_route.dart';
+part 'routes/modular_route.dart';
+part 'routes/id_route.dart';
+part 'routes/optional_id_route.dart';
 
 part 'exceptions/page_not_found.dart';
 part 'exceptions/route_not_found.dart';