+++ lib/navigator/modular_history.dart
@@ -2,7 +2,7 @@ part of modular_navigation;
 
 class ModularHistory<TPageParameters extends PageParameters,
     TModularPage extends ModularPage<TPageParameters>> {
-  final ModularRoute route;
+  final BaseModularRoute route;
   final TModularPage page;
   const ModularHistory(this.route, this.page);