+++ lib/routes/modular_route.dart
@@ -1,63 +1,41 @@
 part of modular_navigation;
 
 class ModularRoute<TPageParameters extends PageParameters,
-    TModularPage extends ModularPage<TPageParameters>> {
-  final BaseModule module;
-  final String _route;
-  final TModularPage Function(Map<String, String?> parameters) createPage;
-  final FutureOr<bool> Function(
-      ModularHistory route, ModularRouterDelegate delegate)? guard;
+        TModularPage extends ModularPage<TPageParameters>>
+    extends BaseModularRoute<TPageParameters, TModularPage> {
+  ModularRoute(
+      {required BaseModule module,
+      required String route,
+      required TModularPage Function(Map<String, String?> params) createPage,
+      FutureOr<bool> Function(
+              ModularHistory route, ModularRouterDelegate delegate)?
+          guard})
+      : super(
+          module: module,
+          route: route,
+          createPage: createPage,
+          guard: guard,
+        );
 
-  ModularRoute({
-    required this.module,
-    required String route,
-    required this.createPage,
-    this.guard,
-  }) : _route = combinePath(module.route, route) {
-    assert(TPageParameters != dynamic && TPageParameters != PageParameters);
-    assert(TModularPage != dynamic &&
-        !(TModularPage is ModularPage<PageParameters>));
+  ModularLink<TPageParameters> createLink(TPageParameters parameters) {
+    return super._createLink(parameters: parameters);
   }
 
-  String get route => _route;
-
-  bool isPageRoute(Type pageType) => pageType == TModularPage;
-
-  ModularPage<TPageParameters> create({
-    TPageParameters? parameters,
-  }) =>
-      createPage(parameters!.map);
-
-  ModularLink createLink({
-    TPageParameters? parameters,
-  }) =>
-      ModularLink(
-        route: this,
-        parameters: parameters!,
-      );
+  TModularPage create(TPageParameters parameters) {
+    return super._create(parameters: parameters) as TModularPage;
+  }
 
   void navigate({
     required BuildContext context,
-    TPageParameters? parameters,
+    required TPageParameters parameters,
     bool clearHistory = false,
     bool removeCurrent = false,
-  }) =>
-      context.routeDelegate.navigateToPage(
-        create(parameters: parameters!),
-        clearHistory: clearHistory,
-        removeCurrent: removeCurrent,
-      );
-
-  @override
-  operator ==(Object o) {
-    if (identical(this, o)) return true;
-
-    return o is ModularRoute &&
-        o.route == route &&
-        o.module == module &&
-        o.guard == guard;
+  }) {
+    super._navigate(
+      context: context,
+      parameters: parameters,
+      clearHistory: clearHistory,
+      removeCurrent: removeCurrent,
+    );
   }
-
-  @override
-  int get hashCode => route.hashCode & module.hashCode & guard.hashCode;
 }