+++ lib/routes/optional_id_route.dart
@@ -1,0 +1,21 @@
+part of modular_navigation;
+
+class OptionalIdPageRoute<
+        TModularPage extends ModularPage<OptionalIdPageParameters>>
+    extends ModularRoute<OptionalIdPageParameters, TModularPage> {
+  OptionalIdPageRoute(
+      {required BaseModule module,
+      required String route,
+      required TModularPage Function(OptionalIdPageParameters params)
+          createPage,
+      FutureOr<bool> Function(
+              ModularHistory route, ModularRouterDelegate delegate)?
+          guard})
+      : super(
+          module: module,
+          route: route,
+          createPage: (params) =>
+              createPage(OptionalIdPageParameters.fromMap(params)),
+          guard: guard,
+        );
+}